Rename repository and stub first notebook
Why these changes are being introduced:

This initial commit renames the repository and stubs an first notebook
to establish some testing and linting conventions with the notebook .py file.

NOTE: this is the first implementation of a notebook with the underlying
'marimo-notebook-template' template.  It is anticipated that many, if not most,
marimo notebook repositories will contain only a single notebook and thus
could use the default notebook.py file setup in the template.

Side effects of this change:
* None

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/IN-1438